% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ktau.to.coppar}
\alias{ktau.to.coppar}
\title{Convert the Kendall's tau into the copula parameter}
\usage{
ktau.to.coppar(ktau, copfam)
}
\arguments{
\item{ktau}{a numeric value that denotes the Kendall's tau.}

\item{copfam}{a character string that denotes the copula family.}
}
\description{
Convert the Kendall's tau into the copula parameter
}
